Uploading individual files
Decided to upload the individual files used within this project for easier viewing. Pictures have been removed from the project for copyright reasons. This changes the CSS formatting of the website, but the general idea still remains. Other images could be added to the project if the links are changed accordingly.